Rename: api 폴더 구조 및 파일명 수정
