+++ src/app/api/activities-api.ts
@@ -1,9 +1,8 @@
-import axios, { Axios, AxiosError, ResponseType } from "axios";
-import { instance } from "@/apis/apis";
+import { instance } from "@/app/api/api";
 import { DataType } from "@/types/activities-type";
 import { ReviewData } from "@/types/activities-type";
 import { AvailableSchedule } from "@/types/activities-type";
-import { requestor } from "@/apis/config/requestor";
+import { requestor } from "@/app/api/config/requestor";
 import { ActivityListResponse } from "@/types/activities-type";
 
 const BASE_URL = "https://sp-globalnomad-api.vercel.app/4-17";
@@ -26,8 +25,8 @@ export const getActivities = async (
 
 // 체험 상세조회
 export async function getDatas(activityId: number): Promise<DataType> {
-  const response = await fetch(`${BASE_URL}/activities/${activityId}`);
-  return await response.json();
+  const response = await instance.get<DataType>(`/activities/${activityId}`);
+  return response.data;
 }
 
 export const getActivityDetails = async (activityId: number) => {
@@ -42,10 +41,10 @@ export async function getAvailableSchedule(
   month: number,
   activityId: number
 ): Promise<AvailableSchedule> {
-  const response = await fetch(
-    `${BASE_URL}/activities/${activityId}/available-schedule`
+  const response = await instance.get<AvailableSchedule>(
+    `/activities/${activityId}/available-schedule`
   );
-  return await response.json();
+  return response.data;
 }
 
 export const getSchedule = async (
@@ -53,17 +52,16 @@ export const getSchedule = async (
   month: string,
   activityId: number
 ): Promise<AvailableSchedule[]> => {
-  const response = await requestor.get(
+  const response = await instance.get<AvailableSchedule[]>(
     `/activities/${activityId}/available-schedule`,
     {
       params: {
-        year: year,
-        month: month,
+        year,
+        month,
       },
     }
   );
-  const schedule = response.data;
-  return schedule;
+  return response.data;
 };
 
 //체험 리뷰 조회
@@ -72,13 +70,14 @@ export async function getReviews(
   page: number,
   size: number
 ): Promise<ReviewData> {
-  const params = new URLSearchParams({
-    activityId: activityId.toString(),
-    page: page.toString(),
-    size: size.toString(),
-  });
-  const response = await fetch(
-    `${BASE_URL}/activities/${activityId}/reviews?${params}`
+  const response = await instance.get<ReviewData>(
+    `/activities/${activityId}/reviews`,
+    {
+      params: {
+        page,
+        size,
+      },
+    }
   );
-  return await response.json();
+  return response.data;
 }
