+++ src/app/layout.tsx
@@ -1,6 +1,5 @@
 import type { Metadata } from "next";
 import "@/styles/globals.css";
-import Head from "next/head";
 
 export const metadata: Metadata = {
   title: "GlobalNomad",
@@ -15,18 +14,6 @@ export default function RootLayout({
 }) {
   return (
     <>
-      <Head>
-        <meta
-          name='description'
-          content='당신의 일상을 특별하게 만드는 한 번의 클릭'
-        />
-        <meta name='og:title' content='GlobalNomad' />
-        <meta
-          name='og:description'
-          content='당신의 일상을 특별하게 만드는 한 번의 클릭'
-        />
-        <meta name='og:image' content='/logo.svg' />
-      </Head>
       <html lang='ko'>
         <body>{children}</body>
       </html>
